refactor(@dpc-sdp/nuxt-ripple-cli): :recycle: refactor favicon generator into cli tool
