+++ packages/nuxt-ripple-cli/src/commands/favicon/lib.d.ts
