+++ packages/nuxt-ripple/package.json
@@ -19,8 +19,6 @@
     "@nuxtjs/robots": "^3.0.0",
     "@vueuse/core": "^9.13.0",
     "change-case": "^4.1.2",
-    "defu": "^6.1.2",
-    "jsonapi-parse": "^2.0.1",
-    "rfg-api": "^0.5.3"
+    "defu": "^6.1.2"
   }
 }
